Issue #931 Use separate package for Restrictions
