+++ api/src/main/java/jakarta/data/metamodel/Attribute.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023,2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023,2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -19,8 +19,8 @@ package jakarta.data.metamodel;
 
 import java.util.Set;
 
-import jakarta.data.Restrict;
-import jakarta.data.Restriction;
+import jakarta.data.restrict.Restrict;
+import jakarta.data.restrict.Restriction;
 
 /**
  * Represents an entity attribute in the {@link StaticMetamodel}.
