+++ api/src/main/java/jakarta/data/metamodel/TextAttribute.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2023,2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2023,2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,9 +17,9 @@
  */
 package jakarta.data.metamodel;
 
-import jakarta.data.Restrict;
 import jakarta.data.Sort;
-import jakarta.data.TextRestriction;
+import jakarta.data.restrict.Restrict;
+import jakarta.data.restrict.TextRestriction;
 
 /**
  * Represents an textual entity attribute in the {@link StaticMetamodel}.
