+++ api/src/test/java/jakarta/data/metamodel/SortableAttributeTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024 Contributors to the Eclipse Foundation
+ * Copyright (c) 2024,2025 Contributors to the Eclipse Foundation
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -17,11 +17,12 @@
  */
 package jakarta.data.metamodel;
 
-import jakarta.data.BasicRestriction;
-import jakarta.data.CompositeRestriction;
-import jakarta.data.Operator;
-import jakarta.data.Restriction;
 import jakarta.data.Sort;
+import jakarta.data.restrict.BasicRestriction;
+import jakarta.data.restrict.CompositeRestriction;
+import jakarta.data.restrict.Operator;
+import jakarta.data.restrict.Restriction;
+
 import org.assertj.core.api.SoftAssertions;
 import org.junit.jupiter.api.Test;
